feat: mode export as json logic to library
