+++ library/lib/index.tsx
@@ -1,7 +1,6 @@
 import ReactDOM from "react-dom/client"
 import { AppWithProvider } from "./App"
 import { ReactFlowInstance, type Node, type Edge } from "@xyflow/react"
-
 export { type Node } from "@xyflow/react"
 
 export class Apollon2 {
@@ -25,6 +24,27 @@ export class Apollon2 {
     }
     return []
   }
+
+  public exportAsJson(diagramName: string) {
+    if (this.reactFlowInstance) {
+      const data = {
+        version: "apollon2",
+        title: diagramName,
+        nodes: this.reactFlowInstance.getNodes(),
+        edges: this.reactFlowInstance.getEdges(),
+      }
+      const jsonString = `data:text/json;chatset=utf-8,${encodeURIComponent(
+        JSON.stringify(data)
+      )}`
+      const link = document.createElement("a")
+      link.href = jsonString
+      link.download = `${diagramName}.json`
+
+      link.click()
+    }
+    return []
+  }
+
   public getEdges(): Edge[] {
     if (this.reactFlowInstance) {
       return this.reactFlowInstance.getEdges()
