+++ standalone/webapp/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@apollon2/library": "0.0.1",
+    "html-to-image": "^1.11.11",
     "react": "18.3.1",
     "react-dom": "18.3.1"
   },
